+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ New features
 * Add TrialDesign class
 * Add :code:`modeling.create_trial_design`
 * Add :code:`modeling.add_arm`
+* Add :code:`modeling.add_observations`
 
 Bugfixes
 ========